requirements.txt
aiohttp==2.0.2
python-chess==0.18.2
CairoSVG==2.0.3
